+++ grails-app/services/org/bbop/apollo/PermissionService.groovy
@@ -583,7 +583,7 @@ class PermissionService {
         PermissionEnum highestEnum = PermissionEnum.NONE
         for(PermissionEnum permissionEnum : permissionEnums){
             if(permissionEnum.rank>highestEnum.rank){
-                highestEnum = permissionEnum ;
+                highestEnum = permissionEnum
             }
         }
         return highestEnum