+++ grails-app/services/org/bbop/apollo/ProxyService.groovy
@@ -5,7 +5,8 @@ import grails.transaction.Transactional
 @Transactional(readOnly = true)
 class ProxyService {
 
-    private final static List<String> defaultProxies = ["http://golr.berkeleybop.org/solr/select"]
+//    private final static List<String> defaultProxies = ["http://golr.berkeleybop.org/"]
+    def grailsApplication
 
     /**
      * Looks through all proxies to return valid proxies
@@ -13,8 +14,38 @@ class ProxyService {
      * @return
      */
     Proxy findProxyForUrl(String referenceUrl){
-        Proxy proxy = Proxy.findByReferenceUrl(referenceUrl)
+        Proxy proxy = Proxy.findByReferenceUrlAndActive(referenceUrl,true)
+        if(!proxy){
+            def proxyList = Proxy.findAllByReferenceUrl(referenceUrl,[sort:"fallbackOrder",order:"asc"])
+            if(proxyList){
+                return proxyList.first()
+            }
+        }
         return proxy
+    }
+
+    @Transactional
+    def initProxies(){
+        def proxies = grailsApplication.config.apollo.proxies
+
+        for(proxyConfig in proxies){
+            def proxy = Proxy.findByReferenceUrlAndTargetUrl(proxyConfig.referenceUrl,proxyConfig.targetUrl)
+
+            if(!proxy){
+                proxy = new Proxy(
+                        referenceUrl: proxyConfig.referenceUrl
+                        , targetUrl: proxyConfig.targetUrl
+                        ,active: proxyConfig.active
+                        ,fallbackOrder: proxyConfig.fallbackOrder
+                ).save(failOnError: false,insert: true)
+            }
+            else
+            if (proxyConfig.replace) {
+                active: proxyConfig.active
+                fallbackOrder: proxyConfig.fallbackOrder
+                proxy.save(failOnError: false,insert: false)
+            }
+        }
     }
 
     /**